requirements.txt
certifi==2018.4.16
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.5.0
Django==2.0.4
django-allauth==0.35.0
django-braces==1.13.0
django-debug-toolbar==1.9.1
django-extensions==2.0.7
django-request-logging==0.6.5
django-rest-auth==0.9.3
django-rest-swagger==2.2.0
djangorestframework==3.8.2
google-api-python-client==1.6.7
httplib2==0.11.3
idna==2.7
itypes==1.1.0
Jinja2==2.10
MarkupSafe==1.0
oauth2client==4.1.2
oauthlib==2.1.0
openapi-codec==1.3.2
pkg-resources==0.0.0
psycopg2==2.7.4
psycopg2-binary==2.7.4
pyasn1==0.4.3
pyasn1-modules==0.2.1
python3-openid==3.1.0
pytz==2018.4
requests==2.19.0
requests-oauthlib==1.0.0
rsa==3.4.2
simplejson==3.15.0
six==1.11.0
sqlparse==0.2.4
uritemplate==3.0.0
urllib3==1.23
